Update ref. count QIR signature
The spec. uses int32_t type for the increment argument.

+++ mlir/qir_qrt/qir-qrt.cpp
@@ -429,12 +429,12 @@ bool __quantum__rt__result_equal(Result *res, Result *comp) {
 Result *__quantum__rt__result_get_one() { return ResultOne; }
 Result *__quantum__rt__result_get_zero() { return ResultZero; }
 
-void __quantum__rt__string_update_reference_count(void *str, int64_t count) {
+void __quantum__rt__string_update_reference_count(void *str, int32_t count) {
   // TODO
   if (verbose) std::cout << "CALL: " << __PRETTY_FUNCTION__ << "\n";
 }
 
-void __quantum__rt__result_update_reference_count(Result *, int64_t count) {
+void __quantum__rt__result_update_reference_count(Result *, int32_t count) {
   // TODO
   if (verbose) std::cout << "CALL: " << __PRETTY_FUNCTION__ << "\n";
 }
